+++ spec/integration/plugin_spec.rb
@@ -2,6 +2,7 @@
 
 require 'timeout'
 require 'net/http'
+require 'socket'
 
 TestTakesTooLongError = Class.new(StandardError)
 
@@ -40,6 +41,7 @@ module Puma
             'workers.spawned_threads' => 1,
             'workers.max_threads' => 1,
             'workers.requests_count' => 0,
+            'workers.busy_threads' => 0,
             'queue.backlog' => 0,
             'queue.capacity' => 1
           }
@@ -74,24 +76,21 @@ module Puma
 
       context 'when dogstatsd target' do
         let(:config) { 'dogstatsd' }
-        let(:expected_telemetry) do
-          %w[
-            workers.booted:1|g
-            workers.total:1|g
-            workers.spawned_threads:1|g
-            workers.max_threads:1|g
-            workers.requests_count:0|g
-            queue.backlog:0|g
-            queue.capacity:1|g
-          ]
-        end
+        let(:hostname) { Socket.gethostname }
 
         it "doesn't crash" do
           true until (line = @server.next_line).include?('DEBUG -- : Statsd')
 
           lines = ([line.slice(/workers.*/)] + Array.new(6) { @server.next_line.strip })
 
-          expect(lines).to eq(expected_telemetry)
+          # workers.busy_threads should have hostname tag, others should not
+          expect(lines[0]).to eq('workers.booted:1|g')
+          expect(lines[1]).to eq('workers.total:1|g')
+          expect(lines[2]).to eq('workers.spawned_threads:1|g')
+          expect(lines[3]).to eq('workers.max_threads:1|g')
+          expect(lines[4]).to eq('workers.requests_count:0|g')
+          expect(lines[5]).to eq("workers.busy_threads:0|g|#process:#{hostname}")
+          expect(lines[6]).to eq('queue.backlog:0|g')
         end
       end